Select the parent with which to view the diff
Users can now press ctrl key + parent hash (from the details view) to
switch the commit we see the diff with. The parents are presented on
separate lines, and the chosen parent is in bold.

This is done by saving the current parent we view the diff with in the
frontend and each time we request commit details we send the parent
we want to diff with.